backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-multipart==0.0.22
python-dotenv==1.1.0
sqlalchemy==2.0.48
bcrypt==5.0.0
python-jose[cryptography]==3.5.0
email-validator==2.3.0
itsdangerous==2.2.0
openai==1.101.0
pymupdf==1.26.4
PyPDF2==3.0.1
tiktoken==0.11.0
pyyaml==6.0.2
